Switch to async generator
