+++ xml/mod.ts
@@ -82,5 +82,5 @@
 export * from "./types.ts";
 export * from "./parse_stream.ts";
 export * from "./parse.ts";
-export * from "./record_stream.ts";
+export * from "./parse_records.ts";
 export * from "./stringify.ts";
